On User Delete action check if logged user try to delete himself.

+++ src/Controllers/UserController.php
@@ -218,6 +218,18 @@ class UserController extends Controller
         //$id = $this->decode($hash);
         $user = $this->userRepository->findById($id);
 
+        // Check to be sure user cannot delete himself
+		if(Sentinel::getUser()->id == $user->id)
+		{
+			$message = "You cannot remove yourself!";
+
+			if($request->ajax()) {
+				return response()->json($message, 422);
+			}
+			session()->flash('error', $message);
+			return redirect()->route('users.index');
+		
+
         // Remove the user
         $user->delete();
